// src/app/template-driven/user.interface.ts
import { Theme } from './theme.interface';

export interface User {
    username: string; // text
    email: string; // text
    password: string; // text
    age?: number; // number
    gender?: string; // radio
    role?: string; // select (primitive)
    theme?: Theme; // select (object)
    topics?: string[]; // multiple select
    isActive?: boolean; // checkbox
    toggle?: string; // checkbox toggle either 'toggled' or 'untoggled'
}